+++ src/command/uninstall.rs
@@ -20,7 +20,7 @@ impl Command for Uninstall {
         let version = VersionSpec::default();
         let tool = Spec::from_str_and_version(&self.tool, version);
 
-        tool.resolve(session)?.uninstall(session)?;
+        tool.uninstall()?;
 
         session.add_event_end(ActivityKind::Uninstall, ExitCode::Success);
         Ok(ExitCode::Success)
